src/requirements.txt
tensorflow-gpu==1.3
keras==2.2.0
pandas==0.23.3
psutil==5.4.6
tqdm==4.23.4
more_itertools==4.2.0
bs4==0.0.1
lxml==4.2.3
jupyter==1.0.0